[Chore 157992786] Improve on PEP8 and restructure models
