+++ API/data_store/collections.py
@@ -1,22 +1,24 @@
+"""The data structures to store non-persistent data in models"""
 import re
 from datetime import datetime
-
 from passlib.handlers.bcrypt import bcrypt
-
-from models.base import BaseModel
-from models.user import User
+from app.models import BaseModel, User
 
 
 class NonPersistentCollection:
+    """The base data structure, with functions to transact models"""
 
     def __init__(self):
+        """Creates a data structure and index for item ids"""
         self.data = {}
         self.index = 1
 
     def query_all(self):
+        """Get all items in this collection"""
         return self.data
 
     def query_all_where_field_eq(self, field, value):
+        """Query all items by a specific field"""
         result = []
         for item in self.data.values():
             if item.to_json_object()[field] == value:
@@ -25,6 +27,7 @@ class NonPersistentCollection:
         return result
 
     def insert(self, item):
+        """Insert an item into the collection"""
         assert isinstance(item, BaseModel)
         item.created_at = datetime.now()
         self.data[self.index] = item
@@ -32,32 +35,38 @@ class NonPersistentCollection:
         self.index += 1
 
     def set(self, item, item_id):
+        """Updates an item fields"""
         assert isinstance(item, BaseModel)
         item.updated_at = datetime.now()
         self.data[item_id] = item
 
     def query(self, item_id):
+        """Query an item by id"""
         return self.data.get(item_id)
 
     def query_by_field(self, field, value):
+        """Query an item by a specific field"""
         for item in self.data.values():
             if item.to_json_object()[field] == value:
                 return item
 
     def delete(self, item_id):
+        """Delete an item from the collection"""
         del self.data[item_id]
 
     def is_valid(self, item):
+        """Override this method to check if an item is valid or not"""
         return True, []
 
     def clear(self):
+        """Clear the items in this collection"""
         self.data = {}
 
 
 class UserCollection(NonPersistentCollection):
 
     def insert(self, item):
-        # encrypt the password
+        """Encrypt password before adding user into the collection"""
         assert isinstance(item, User)
         item.password = bcrypt.encrypt(item.password)
 
@@ -65,6 +74,7 @@ class UserCollection(NonPersistentCollection):
         super().insert(item)
 
     def is_valid(self, item):
+        """Check if the response has valid user details"""
         errors = {}
         if not item.get("firstname"):
             errors['firstname'] = "First name is required"
@@ -93,7 +103,9 @@ class UserCollection(NonPersistentCollection):
 
 
 class RequestCollection(NonPersistentCollection):
+
     def is_valid(self, item):
+        """Check if a request has valid fields"""
         errors = {}
         if not item.get("product_name"):
             errors["product_name"] = "Product name must be provided"
@@ -106,6 +118,7 @@ class RequestCollection(NonPersistentCollection):
 
 class NotificationCollection(NonPersistentCollection):
     def is_valid(self, item):
+        """Check if a notification has valid fields"""
         errors = {}
         if not item.get("message"):
             errors["message"] = "Notification message must be provided"
@@ -115,6 +128,7 @@ class NotificationCollection(NonPersistentCollection):
 
 class FeedbackCollection(NonPersistentCollection):
     def is_valid(self, item):
+        """Check whether a feedback object has valid fields"""
         errors = {}
         if not item.get("message"):
             errors["message"] = "Feedback message must be provided"
