+++ API/tests/test_sign_up.py
@@ -1,5 +1,5 @@
 from tests.base_test import BaseTestCase
-from models.user import User
+from app.models import User
 import json
